Move resourceView to be sibling of search, not subdirectory
