+++ plugin-hrm-form/src/components/resources/resourceView/ExpandableSection.tsx
@@ -17,8 +17,8 @@ import React, { useState } from 'react';
 import ArrowRightTwoTone from '@material-ui/icons/ArrowRightTwoTone';
 import ArrowDropDownTwoTone from '@material-ui/icons/ArrowDropDownTwoTone';
 
-import { SectionCollapse } from '../../../../styles/search';
-import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../../../../styles/ReferrableResources';
+import { SectionCollapse } from '../../../styles/search';
+import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../../../styles/ReferrableResources';
 
 type ExpandableSectionProps = {
   title: string | JSX.Element;
